Try without kubeflow example
